requirements.txt
flask
binance
python-binance
ta-lib
pandas
numpy
scikit-learn
xgboost
ta
joblib
requests
openai
google-generativeai